コードの整理

+++ app/Services/Admin/ContactForm/UpdateService.php
@@ -9,6 +9,7 @@ use App\Domain\Repositories\ContactForm\ContactFormRepository;
 use App\Enums\PhotoType;
 use App\Http\Requests\Admin\ContactForm\UpdateRequest;
 use App\Services\BaseService;
+use Illuminate\Support\Facades\Storage;
 
 class UpdateService extends BaseService
 {
@@ -53,22 +54,40 @@ class UpdateService extends BaseService
             $model
         );
 
-        // お問い合わせ画像テーブルを登録（Delete→Insert）
+        // お問い合わせ画像テーブルを登録（差分チェックして更新）
         $contactFormImages = $this->contactFormImageRepository->getByContactFormId($contactFormId);
-        foreach ($contactFormImages as $contactFormImage) {
-            $this->contactFormImageRepository->delete($contactFormImage->id);
+        foreach ([$request->delete_image_1, $request->delete_image_2, $request->delete_image_3] as $i => $isDelete) {
+            if ('1' !== $isDelete) {
+                continue;
+            }
+            $this->contactFormImageRepository->delete($contactFormImages[$i]->id);
+            Storage::delete(PhotoType::Contact->dirName() . '/' . $contactFormImages[$i]->file_name);
         }
-        foreach ($request['image_files'] as $i => $imageFile) {
-            $fileName = $imageFile->getClientOriginalName();
-            $this->contactFormImageRepository->create(
-                [
+
+        foreach ([$request->image_file_1, $request->image_file_2, $request->image_file_3] as $i => $imageFile) {
+            $contactFormImage = $contactFormImages[$i] ?? null;
+            if (is_null($imageFile)) {
+                continue;
+            }
+            $newFileName = $imageFile->getClientOriginalName();
+
+            // 変更がある場合のみ削除と再登録
+            if (!$contactFormImage || $contactFormImage->file_name !== $newFileName) {
+                if ($contactFormImage) {
+                    $this->contactFormImageRepository->delete($contactFormImage->id);
+                    Storage::delete(PhotoType::Contact->dirName() . '/' . $contactFormImage->file_name);
+                }
+
+                $this->contactFormImageRepository->create([
                     'contact_form_id' => $contactFormId,
-                    'file_name' => $fileName,
-                ]
-            );
-            //s3に画像をアップロード
-            $imageFile->storeAs(PhotoType::Contact->dirName() . '/', $fileName);
+                    'file_name' => $newFileName,
+                ]);
+
+                // s3に画像をアップロード
+                $imageFile->storeAs(PhotoType::Contact->dirName() . '/', $newFileName);
+            }
         }
+
 
         return $contactForm;
     }
